backend/requirements.txt
falcon==4.0.2
python-dotenv==1.1.1
SQLAlchemy==2.0.41
psycopg2-binary==2.9.10